Update chap02
The first line edited is self explanatory, that is a typo. If we create a sawtooth signal, we should say "sawtooth" instead of "triangle"

The second edit is needed, since the spectrum that we are modifying is from line 440, which is before creating the new sawtooth signal. 
I discovered it by making changes to duration and framerate of the sawtooth signal, and when playing the two comparative audios (with just different phases) the lenghts of the audios were completely different.